AddData moved to src/pages
